use shared GameTime and update agent guide
